Better fix for #200. frame_size is the only variable that need to be in 64bit

+++ src/algos/demosaicing.h
@@ -10,7 +10,7 @@ WORD *debayer_buffer(WORD *buf, int *width, int *height,
 		interpolation_method interpolation, sensor_pattern pattern);
 int debayer(fits*, interpolation_method);
 void get_debayer_area(const rectangle *area, rectangle *debayer_area,
-		const rectangle *image_area, int64_t *debayer_offset_x,
-		int64_t *debayer_offset_y);
+		const rectangle *image_area, int *debayer_offset_x,
+		int *debayer_offset_y);
 
 #endif
